requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.2.2
numpy<2.0
scikit-learn==1.4.1.post1
tqdm==4.66.1
sentence-transformers==2.7.0
requests==2.31.0
groq==0.9.0
huggingface_hub==0.23.1
transformers==4.40.0
streamlit==1.28.1